master

+++ include/botfs_config.h
@@ -38,7 +38,7 @@
 #define BOTFS_HEADER_POS	1			/**< Blockadresse des Volume-Headers (erster Block ist Header der Volume-Header-Datei) */
 #define BOTFS_BLOCK_SIZE	512U		/**< Groesse eines Blocks (kleinste adressierbare Einheit) in Byte */
 #define BOTFS_ROOT_SIZE		UINT16_MAX	/**< Groesse des Root-Verzeichnisses in Byte */
-#define BOTFS_MAX_FILENAME	(128UL - (sizeof(botfs_file_descr_t) + 1))		/**< max. Dateinamenlaenge in Zeichen (116 Zeichen + 0-Byte + 11 Byte Dateidescr. = 128 Byte Daten pro Datei) */
+#define BOTFS_MAX_FILENAME	(128UL - (uint32_t)(sizeof(botfs_file_descr_t) + 1)) /**< max. Dateinamenlaenge in Zeichen (116 Zeichen + 0-Byte + 11 Byte Dateidescr. = 128 Byte Daten pro Datei) */
 #define BOTFS_MAX_VOLUME_SIZE ((uint32_t) UINT16_MAX * BOTFS_BLOCK_SIZE)	/**< max. Groesse des Volumes in Byte */
 #define BOTFS_MAX_FILE_SIZE	((uint32_t) INT16_MAX * BOTFS_BLOCK_SIZE)		/**< max. Groesse einer Datei in Byte */
 #define BOTFS_DIR_BLOCK_CNT	(BOTFS_ROOT_SIZE / BOTFS_BLOCK_SIZE)			/**< Anzahl der Root-Dir-Bloecke */
